requirements.txt
uproot
awkward
numpy
pandas
matplotlib
hist
boost-histogram
dask[distributed]
pyyaml
pytest
jupyterlab
